% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybinombarplot.R
\name{mybin}
\alias{mybin}
\title{Binomial Barplot}
\usage{
mybin(iter = 10000, n = 10, p = 0.7)
}
\arguments{
\item{iter}{number of times the barplot is created or made}

\item{n}{number of bars produced}

\item{p}{value of probability}
}
\value{
a binomial barplot
}
\description{
Binomial Barplot
}
\examples{
mybin(iter=100,n=10, p=0.7)
}
